requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
python-dotenv

